+++ src/game/ItemHandler.cpp
@@ -38,7 +38,8 @@ void WorldSession::HandleSwapInvItemOpcode( WorldPacket & recv_data )
 {
     WorldPacket data;
 	UpdateData upd;
-    uint8 srcslot, dstslot;
+///    uint8 srcslot, dstslot;
+    uint32 srcslot, dstslot;
 
     recv_data >> srcslot >> dstslot;
 
@@ -59,7 +60,7 @@ void WorldSession::HandleSwapInvItemOpcode( WorldPacket & recv_data )
 
     // check to make sure items are not being put in wrong spots
 	if ( (srcslot >= INVENTORY_SLOT_BAG_START && srcslot < BANK_SLOT_BAG_END)&&
-		 (dstslot >= EQUIPMENT_SLOT_START && dstslot < EQUIPMENT_SLOT_END)
+	 (dstslot >= EQUIPMENT_SLOT_START && dstslot < EQUIPMENT_SLOT_END)
 	   )
     {
         uint8 error = GetPlayer()->CanEquipItemInSlot(dstslot, srcitem->GetProto());
